docs: Add Javadoc and comments for multithreaded solving
