+++ core/src/main/java/ai/timefold/solver/enterprise/core/multithreaded/MultiThreadedLocalSearchDecider.java
@@ -111,14 +111,19 @@ final class MultiThreadedLocalSearchDecider<Solution_> extends LocalSearchDecide
     @Override
     public void phaseEnded(LocalSearchPhaseScope<Solution_> phaseScope) {
         super.phaseEnded(phaseScope);
-        // Tell the move thread runners to stop
-        // Don't clear the operationsQueue to avoid moveThreadBarrier deadlock:
-        // The MoveEvaluationOperations are already cleared and the new ApplyStepOperation isn't added yet.
         DestroyOperation<Solution_> destroyOperation = new DestroyOperation<>();
+        // Set to -2, since unless step index overflows, no stepIndex in MoveThreadRunner
+        // will be equal to it (it is set to -1 initially, so cannot use that).
         nextSynchronizedOperationIndex.set(-2);
         nextSynchronizedOperation.set(destroyOperation);
+
+        // Unblock the Move Threads, so they will get the DestroyOperation
         resultQueue.endPhase();
+
+        // Wait for all Move Threads to consume the DestroyOperation
         shutdownMoveThreads();
+
+        // Update the scoreCalculationCount
         long childThreadsScoreCalculationCount = 0;
         for (MoveThreadRunner<Solution_, ?> moveThreadRunner : moveThreadRunnerList) {
             childThreadsScoreCalculationCount += moveThreadRunner.getCalculationCount();
@@ -152,14 +157,33 @@ final class MultiThreadedLocalSearchDecider<Solution_> extends LocalSearchDecide
         int stepIndex = stepScope.getStepIndex();
 
         int selectMoveIndex = 0;
-        AtomicLong hasNextRemaining = new AtomicLong(moveSelectorList.size());
+
+        // We will have moveThreadCount == moveSelectorList.size() distinct iterators
+        AtomicLong hasNextRemaining = new AtomicLong(moveThreadCount);
         for (int i = 0; i < moveThreadCount; i++) {
-            SharedNeverEndingMoveGenerator<Solution_> sharedGenerator = new SharedNeverEndingMoveGenerator<>(hasNextRemaining,
+            // Use offset i, increment moveThreadCount so the indices generated by each NeverEndingMoveGenerator
+            // are unique
+            // For instance, for thread count = 4, the sequences generated would be:
+            // [0, 4, 8, ...]
+            // [1, 5, 9, ...]
+            // [2, 6, 10, ...]
+            // [3, 7, 11, ...]
+            NeverEndingMoveGenerator<Solution_> sharedGenerator = new NeverEndingMoveGenerator<>(hasNextRemaining,
                     resultQueue, moveSelectorList.get(i).iterator(), new AtomicBoolean(true), i, moveThreadCount);
             iteratorReference.set(i, sharedGenerator);
         }
         moveIndex.set(0);
+
+        // All variables set up, unblock move generators
         resultQueue.startNextStep(stepIndex);
+
+        // moveIndex = number of generated moves so far
+        // selectMoveIndex = number of moves foraged so far
+        // if selectMoveIndex < moveIndex, then we still want to forage
+        // even if the resultQueue is blocking (as there are moves waiting
+        // in the resultQueue that we haven't consumed).
+        // If the resultQueue is blocking, that mean the iterator is exhausted
+        // and no more moves will come in
         while (selectMoveIndex < moveIndex.get() || !resultQueue.checkIfBlocking()) {
             if (forageResult(stepScope, stepIndex)) {
                 break;
@@ -171,7 +195,8 @@ final class MultiThreadedLocalSearchDecider<Solution_> extends LocalSearchDecide
         resultQueue.blockMoveThreads();
 
         pickMove(stepScope);
-        // Start doing the step on every move thread. Don't wait for the stepEnded() event.
+
+        // Wait for all threads to finish evaluating moves
         resultQueue.deciderSyncOnEnd();
 
         // Set Random state to a deterministic value
@@ -203,6 +228,7 @@ final class MultiThreadedLocalSearchDecider<Solution_> extends LocalSearchDecide
             return true;
         }
         if (result == null) {
+            // The iterator is exhausted
             stepScope.getPhaseScope().getSolverScope().checkYielding();
             return termination.isPhaseTerminated(stepScope.getPhaseScope());
         }
